Finished MightyMouse Images
Overall, I really liked the way this mouse fits in my hand. I really like that my whole hand fits on it and that my thumb can rest on the left side. I think that the back could have been more rounded coming around to the spot for the thumb. I also think that I should have spent a little bit longer sanding to remove the lines in the top of my paint job. And I just have made the pieces that click the buttons bigger so that it is actually functional to click.